# Small Tensor.cat optimization and reformating