Previous implementation uses Tensor.slice which uses pad + shrink. Previous implementation used slice instead of pad so I converted the slice to pad in new one. We do not need shrink in the cat operation.

Hope it is not Chesterton's Fence.

Just a little bit optimization effect on llama but speed is volatile so practically less function call.

This branch
```
using <class 'tinygrad.runtime.ops_metal.MetalCodegen'>
testing llama python run time
built model
assigned empty tensors, doing warmup
codegen         mean runtime:  147.83ms, runs:   162.49,  170.76,  143.67,  145.73,  141.18,  143.99,  141.86,  142.29,  143.80,  142.52
methodcache     mean runtime:  135.79ms, runs:   158.25,  132.46,  133.65,  134.84,  133.17,  133.09,  132.98,  133.04,  133.33,  133.12
profile         mean runtime:  569.00ms, runs:   624.74,  556.08,  556.12,  584.91,  555.04,  560.35,  561.12,  562.85,  566.95,  561.80
         7783971 function calls (7322683 primitive calls) in 5672.633 seconds
``` 

Master
```
using <class 'tinygrad.runtime.ops_metal.MetalCodegen'>
testing llama python run time
built model
assigned empty tensors, doing warmup
codegen         mean runtime:  149.85ms, runs:   167.50,  181.94,  144.61,  143.15,  142.93,  143.16,  143.04,  144.19,  144.28,  143.74
methodcache     mean runtime:  138.88ms, runs:   166.52,  136.16,  135.52,  134.37,  135.71,  134.95,  134.44,  135.37,  138.87,  136.85
profile         mean runtime:  589.25ms, runs:   646.48,  571.32,  572.55,  604.68,  580.91,  584.71,  576.80,  582.33,  589.80,  582.95
         7852323 function calls (7391035 primitive calls) in 5874.847 seconds
```