@tinyb0t 2023-07-26
Changes made in `tinygrad/`:
```
------------------------------------------------------------
files                             insertions       deletions
------------------------------------------------------------
tinygrad/tensor.py                         8               7
------------------------------------------------------------
lines added in the tinygrad folder: 1
```