@wozeparrot 2023-07-26
Can you post before and after results of the speed test you added?