@ZenginU 2023-07-26
Speed fluctuates based on run, my main intention is to convert slice to pad. Theoratically I assumes pad + shrink -> pad can gain us some little speed

Before
```
cat_0                            256x  256    0.03 ms (    3.95 GFLOPS    31.62 GB/s) in torch,    0.41 ms (    0.32 GFLOPS     2.57 GB/s) in tinygrad,   12.29x slower       0.13 MOPS     1.05 MB
cat_1                            256x  256    0.03 ms (    4.24 GFLOPS    33.92 GB/s) in torch,    0.42 ms (    0.32 GFLOPS     2.52 GB/s) in tinygrad,   13.44x slower       0.13 MOPS     1.05 MB
```
Other run Before
```
cat_0                            256x  256    0.03 ms (    4.39 GFLOPS    35.10 GB/s) in torch,    0.37 ms (    0.35 GFLOPS     2.83 GB/s) in tinygrad,   12.42x slower       0.13 MOPS     1.05 MB
cat_1                            256x  256    0.03 ms (    4.38 GFLOPS    35.00 GB/s) in torch,    0.35 ms (    0.38 GFLOPS     3.01 GB/s) in tinygrad,   11.61x slower       0.13 MOPS     1.05 MB
```

After 
```
cat_0                            256x  256    0.03 ms (    4.53 GFLOPS    36.26 GB/s) in torch,    0.37 ms (    0.35 GFLOPS     2.83 GB/s) in tinygrad,   12.80x slower       0.13 MOPS     1.05 MB
cat_1                            256x  256    0.03 ms (    4.55 GFLOPS    36.42 GB/s) in torch,    0.39 ms (    0.34 GFLOPS     2.70 GB/s) in tinygrad,   13.46x slower       0.13 MOPS     1.05 MB
```
Other run after
```
cat_0                            256x  256    0.03 ms (    4.32 GFLOPS    34.52 GB/s) in torch,    0.38 ms (    0.34 GFLOPS     2.74 GB/s) in tinygrad,   12.59x slower       0.13 MOPS     1.05 MB
cat_1                            256x  256    0.03 ms (    3.94 GFLOPS    31.54 GB/s) in torch,    0.41 ms (    0.32 GFLOPS     2.56 GB/s) in tinygrad,   12.30x slower       0.13 MOPS     1.05 MB
```