@wozeparrot 2023-07-26
meh, didn't really expect a speedup.