@ZenginU 2023-07-26
Yeah, main motivation is to get rid of the useless .slice :/